🤖 Refactor(client): group svg 리팩토링
